requirements.txt
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
boto3==1.24.46
botocore==1.27.46
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
h11==0.13.0
idna==3.3
jmespath==1.0.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-csv==0.0.13
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.11
websockets==10.3
wsproto==1.1.0
PyYAML==6.0
